packagegroup-core-boot: make init-ifupdown package a recommendation
init-ifupdown provides an /etc/network/interfaces, which disables
interface management with networkmanager for example. If a network
manager (such as networkmanager) is provided, there may not be a need
for network related init scripts provided by init-ifupdown, so let's
make it a recommendation so it can be easily removed in image recipes
via the NO_RECOMMENDATIONS/BAD_RECOMMENDATIONS mechanism.

Cc: Quentin Schulz <foss+yocto@0leil.net>
Signed-off-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>